chore(lint): expand ruff ruleset (RUF/PERF/PTH/S/G/LOG/C4/TRY/N) and fix findings
Moves the project from the bare E/F/I/W/UP/B/SIM ruff selection to a
broader surface that catches common real-world issues: resource handling
(PTH), logging anti-patterns (G/LOG/TRY400), perf foot-guns (PERF),
security smells (S), naming (N), comprehension/collection style (C4),
unicode confusables (RUF002/003), and miscellaneous RUF correctness.

Extensions:
- TRY003 ignored (verbose raise messages are a feature here, not a bug).
- TRY301 ignored (intentional re-raise inside try for shared except paths).
- tests/**/*.py get S101 (asserts are fine in tests).

Findings resolved:

- S608 (17 sites): all audited; each now carries an explicit "Safe: ..."
  comment + noqa: S608 justification. No site was actually user-controlled.
  One test helper (tests/test_finance_read_api.py) was refactored from a
  dynamic INSERT to two static INSERTs so nothing needs suppression.
- S101 in minx_mcp/: converted production `assert` statements to explicit
  RuntimeError raises so they survive `python -O`.
- S603 (subprocess): added noqa with justification at the 2 sites that
  call internal tools with controlled argv.
- TRY300 / TRY400 / G201: happy-path returns moved to `else`; logger.error
  in except blocks replaced with logger.exception.
- N818: `_SkipNote` -> `_SkipNoteError` plus call-site + test updates.
- PTH123: vault_writer.py lock-file open uses `Path.open(...)`. Stale
  `# noqa: SIM115` on the same line removed (RUF100).
- PTH201, RUF046, RUF059: trivial autofixes.
- RUF002/003: replaced ambiguous unicode dash / minus with ASCII `-`.
- RUF022: `__all__` sorted in a few modules.
- PERF401: 5 `for ... list.append` loops rewritten as comprehensions /
  list.extend where semantics are identical.

Test suite: 946 passed (filterwarnings=error gate active from the prior
P0 commit remains clean).

Made-with: Cursor

+++ minx_mcp/meals/nutrition.py
@@ -12,7 +12,7 @@ ACTIVITY_MULTIPLIERS: dict[str, float] = {
 MIN_CALORIE_FLOOR_KCAL = 1200
 
 # Goal multipliers apply to TDEE before subtracting ``calorie_deficit_kcal``:
-#   cut (−15% vs TDEE): 0.85
+#   cut (-15% vs TDEE): 0.85
 #   maintenance (0%): 1.0
 #   bulk (+10% vs TDEE): 1.10
 GOAL_CALORIE_MULTIPLIERS: dict[str, float] = {
@@ -54,7 +54,7 @@ def calculate_nutrition_targets(
     bmr = round(10.0 * weight_kg + 6.25 * height_cm - 5.0 * age_years + float(SEX_BMR_OFFSETS[sex]))
     tdee = round(float(bmr) * ACTIVITY_MULTIPLIERS[activity_level])
     multiplier = GOAL_CALORIE_MULTIPLIERS[goal]
-    adjusted_tdee = int(round(float(tdee) * multiplier))
+    adjusted_tdee = round(float(tdee) * multiplier)
     calorie_target = max(adjusted_tdee - calorie_deficit_kcal, MIN_CALORIE_FLOOR_KCAL)
     protein_grams = max(round(weight_kg * protein_g_per_kg), 0)
     fat_grams = max(round(weight_kg * fat_g_per_kg), 0)
